+++ practice-app/.env.example
@@ -1,3 +1,3 @@
-DB_USERNAME=postgres
-DB_PASSWORD=Practice1.
-SECRET_KEY=sadaskdaksdnmask
+DB_USERNAME=
+DB_PASSWORD=
+SECRET_KEY=
