+++ practice-app/app/__init__.py
@@ -1,4 +1,3 @@
-from .views import app
 from dotenv import load_dotenv
-
 load_dotenv()
+from .views import app
