+++ practice-app/requirements.txt
@@ -2,21 +2,31 @@ blinker==1.6.2
 certifi==2022.12.7
 charset-normalizer==3.1.0
 click==8.1.3
+dominate==2.7.0
+exceptiongroup==1.1.1
 Flask==2.3.1
 Flask-Bootstrap==3.3.7.1
 Flask-Login==0.6.2
 Flask-SQLAlchemy==3.0.3
 Flask-WTF==1.1.1
+greenlet==2.0.2
 idna==3.4
 importlib-metadata==6.6.0
+iniconfig==2.0.0
 itsdangerous==2.1.2
 Jinja2==3.1.2
 MarkupSafe==2.1.2
-python-dotenv==1.0.0
+packaging==23.1
+pluggy==1.0.0
 psycopg2-binary==2.9.6
+pytest==7.3.1
+python-dotenv==1.0.0
 requests==2.29.0
 SQLAlchemy==2.0.12
+tomli==2.0.1
+typing_extensions==4.5.0
 urllib3==1.26.15
+visitor==0.1.3
 Werkzeug==2.3.2
 WTForms==3.0.1
 zipp==3.15.0
